+++ common/src/main/java/mca/entity/VillagerLike.java
@@ -133,6 +133,14 @@ public interface VillagerLike<E extends Entity & VillagerLike<E>> extends CTrack
         }
     }
 
+    default boolean canBeAttractedTo(VillagerLike<?> other) {
+        return getTraits().hasSameTrait(Traits.Trait.BISEXUAL, other) || getGenetics().getGender().isMutuallyAttracted(other.getGenetics().getGender());
+    }
+
+    default boolean canBeAttractedTo(PlayerSaveData other) {
+        return getTraits().hasTrait(Traits.Trait.BISEXUAL) || getGenetics().getGender().isMutuallyAttracted(other.getGender());
+    }
+
     default Hand getDominantHand() {
         return getTraits().hasTrait(Traits.Trait.LEFT_HANDED) ? Hand.OFF_HAND : Hand.MAIN_HAND;
     }